requirements.txt
kivy
pillow
zeroconf
git+http://github.com/bitcraft/shutter.git
pysdl2
flask
natsort
netifaces
